test(#3664): fold the foreign-agent warning rows into the install-regressions suite
